Get instr_generator to build
