+++ src/behavior/visualize.rs
@@ -302,20 +302,6 @@ impl BehaviorVisitor<()> for Visualizer<'_> {
         }
     }
 
-    fn visit_emit_global_stmts(&mut self, node: &TreeNode) {
-        if let TreeNode::Action {
-            id,
-            ty: ActionType::EmitGlobalStmts,
-            parent,
-        } = node
-        {
-            self.emit_special_action_node(id, "EmitGlobalStmts");
-            self.emit_edge(parent, id);
-        } else {
-            unreachable!()
-        }
-    }
-
     fn visit_emit_pred(&mut self, node: &TreeNode) {
         if let TreeNode::Action {
             id,
