+++ src/generator.rs
@@ -1,5 +1,5 @@
 pub mod init_generator;
-// pub mod instr_generator;
+pub mod instr_generator;
 pub mod types;
 
 #[cfg(test)]
